% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotArea.R
\name{PlotArea}
\alias{PlotArea}
\title{Plot Area (from GROMACS) by ggplot2}
\usage{
PlotArea(data)
}
\arguments{
\item{data}{an object of class xvg from read_xvg function.\code{\link[ggbond]{read_xvg}}}
}
\value{
a ggplot2 object
}
\description{
Plot Area (from GROMACS) by ggplot2
}
\examples{
{
PlotArea(data)
}

}
