% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotGyrate.R
\name{PlotGyrate}
\alias{PlotGyrate}
\title{Plot Gyrate (from GROMACS) by ggplot2}
\usage{
PlotGyrate(data, legend_names = c("Rg", "Rg_X", "Rg_Y", "Rg_Z"))
}
\arguments{
\item{data}{an object of class xvg from read_xvg function.\code{\link[ggbond]{read_xvg}}}
}
\value{
a ggplot2 object
}
\description{
Plot Gyrate (from GROMACS) by ggplot2
}
\examples{
{
PlotGyrate(data)
}

}
