% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotHbonds.R
\name{PlotHbonds}
\alias{PlotHbonds}
\title{Plot Hbonds (from GROMACS) by ggplot2}
\usage{
PlotHbonds(data, legend_names = c("Hydrogen bonds", "Pairs within 0.35 nm"))
}
\arguments{
\item{data}{an object of class xvg from read_xvg function.\code{\link[ggbond]{read_xvg}}}
}
\value{
a ggplot2 object
}
\description{
Plot Hbonds (from GROMACS) by ggplot2
}
\examples{
{
PlotHbonds(data)
}

}
