% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotRMSD.R
\name{PlotRMSD}
\alias{PlotRMSD}
\title{Plot RMSD (from GROMACS) by ggplot2}
\usage{
PlotRMSD(data)
}
\arguments{
\item{data}{an object of class xvg from read_xvg function.\code{\link[ggbond]{read_xvg}}}
}
\value{
a ggplot2 object
}
\description{
Draw a RMSD plot from GROMACS by ggplot2
}
\examples{
\donttest{
PlotRMSD(datat
}

}
