% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotRMSF.R
\name{PlotRMSF}
\alias{PlotRMSF}
\title{Plot RMSF (from GROMACS) by ggplot2}
\usage{
PlotRMSF(data)
}
\arguments{
\item{data}{an object of class xvg from read_xvg function.\code{\link[ggbond]{read_xvg}}}
}
\value{
a ggplot2 object
}
\description{
Plot RMSF (from GROMACS) by ggplot2
}
\examples{
{
PlotRMSF(data)
}

}
