% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggbond_object.R
\name{print.ggbond}
\alias{print.ggbond}
\title{Print a ggbond layout object}
\usage{
\method{print}{ggbond}(x, ...)
}
\arguments{
\item{x}{A \code{ggbond} object.}

\item{...}{Unused.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a ggbond layout object
}
