% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_xvg.R
\name{read_xvg}
\alias{read_xvg}
\title{read xvg (from GROMACS)}
\usage{
read_xvg(file_path, col_names)
}
\arguments{
\item{file_path}{The path of the xvg file}

\item{col_names}{column names. (If not provided, the default column names will be used.)}
}
\value{
a list
}
\description{
read xvg (from GROMACS)
}
\examples{
\donttest{
read_xvg(path)
}

}
