% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_xvg_s.R
\name{read_xvg_s}
\alias{read_xvg_s}
\title{read a xvg (from GROMACS)}
\usage{
read_xvg_s(file_path, col_names)
}
\arguments{
\item{file_path}{The path of the xvg file}

\item{col_names}{column names. (If not provided, the default column names will be used.)}
}
\value{
a data.frame
}
\description{
read a xvg (from GROMACS)
}
\examples{
\donttest{
read_xvg_s(path)
}

}
