% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varimpPlot.R
\name{varimpPlot}
\alias{varimpPlot}
\title{Plot importance by ggplot2}
\usage{
varimpPlot(d, cutoff)
}
\arguments{
\item{d}{an object of class randomForest from importance function.\code{\link[randomForest]{importance}}}

\item{cutoff}{a cutoff value to filter eligible features.}
}
\value{
a ggplot2 object
}
\description{
Draw a plot from importance by ggplot2
}
\examples{
\donttest{
varimpPlot(d, cutoff=3)
}

}
