% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggbond_object.R
\name{read_ggbond_json}
\alias{read_ggbond_json}
\title{Read a ggbond layout object from JSON}
\usage{
read_ggbond_json(file)
}
\arguments{
\item{file}{Input JSON file path produced by \code{\link[=save_ggbond_json]{save_ggbond_json()}}.}
}
\value{
A \code{ggbond} object.
}
\description{
Read a ggbond layout object from JSON
}
